wire optional sumo demo runners to real traci path
