+++ README.md
@@ -178,14 +178,20 @@ Install Python bindings:
 python -m pip install sumolib traci
 ```
 
-Run the offline pre/post demo summaries:
+Set the SUMO binary folder on PATH if needed:
+
+```bash
+set PATH=C:\Program Files (x86)\Eclipse\Sumo\bin;%PATH%
+```
+
+Run the demo summaries:
 
 ```bash
 python -m sumo_demo.run_pre_system --out artifacts/sumo_pre_system.csv
 python -m sumo_demo.run_post_system --out artifacts/sumo_post_system.csv
 ```
 
-This keeps SUMO as an optional validation/demo layer. The live prototype remains the OpenCV/YOLO runtime.
+This keeps SUMO as an optional validation/demo layer. The live prototype remains the OpenCV/YOLO runtime. If SUMO is installed, the same runners can be extended to use TraCI against a real network.
 
 ## Judge Demo Script (60-90 seconds)
